# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Capacity Fpoly",
        "description": "OpenApi Capacity Fpoly<br>\r\n *      ✅NOTE✅\r\n *      Note: Khi sử dụng api authentication v1 cần nhập token bearer phần Authorize\r\n *      Note: Tìm kiếm api theo các tag\r\n *      Note: Thiếu , gặp bug -> chụp ảnh gửi cho BE sửa 🔧\r\n *      ✅END NOTE✅",
        "contact": {
            "email": ""
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/public/contests": {
            "get": {
                "tags": [
                    "Contest"
                ],
                "description": "Description api contests",
                "operationId": "21335560b55dc3b39b979268e1450fa3",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Tìm kiếm ",
                        "required": false
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Lọc theo trạng thái ",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Lọc theo chiều asc hoặc desc ",
                        "required": false
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Các cột cần lọc  ",
                        "required": false
                    },
                    {
                        "name": "major_id",
                        "in": "query",
                        "description": "Lọc theo chuyên ngành   ",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ status: true , data : data }"
                    },
                    "404": {
                        "description": "{ status: false , message : 'Not found' }"
                    }
                }
            }
        },
        "/api/public/contests/demo": {
            "get": {
                "tags": [
                    "Contest",
                    "ContestDemo"
                ],
                "description": "Description api contests demo",
                "operationId": "3b1993493e7ef52e6b5bd19ca6188906",
                "responses": {
                    "200": {
                        "description": "{ status: true , data : data }"
                    },
                    "404": {
                        "description": "{ status: false , message : 'Not found' }"
                    }
                }
            }
        },
        "/api/public/capacity": {
            "get": {
                "tags": [
                    "Capacity"
                ],
                "description": "Description api capacity",
                "operationId": "35885eef783a04e08d43104f04f5887d",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Tìm kiếm ",
                        "required": false
                    },
                    {
                        "name": "qq",
                        "in": "query",
                        "description": "Tìm kiếm tách chuỗi ",
                        "required": false
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Lọc theo trạng thái ",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Lọc theo chiều asc hoặc desc ",
                        "required": false
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Các cột cần lọc  ",
                        "required": false
                    },
                    {
                        "name": "major_id",
                        "in": "query",
                        "description": "Lọc theo chuyên ngành   ",
                        "required": false
                    },
                    {
                        "name": "skill_id",
                        "in": "query",
                        "description": "Lọc theo skill   ",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ status: true , data : data }"
                    },
                    "404": {
                        "description": "{ status: false , message : 'Not found' }"
                    }
                }
            }
        },
        "/api/public/contests/{id}": {
            "get": {
                "tags": [
                    "Contest"
                ],
                "description": "Description api contests",
                "operationId": "1241e1c4c5b972314cf795f2e34d9803",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id cuộc thi ",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ status: true , data : data }"
                    },
                    "404": {
                        "description": "{ status: false , message : 'Not found' }"
                    }
                }
            }
        },
        "/api/public/capacity/{id}": {
            "get": {
                "tags": [
                    "Capacity"
                ],
                "description": "Description api capacity",
                "operationId": "90df66c6dccd12e824748cf907944a9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id đánh giá năng lực  ",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ status: true , data : data }"
                    },
                    "404": {
                        "description": "{ status: false , message : 'Not found' }"
                    }
                }
            }
        },
        "/api/public/contests/{id}/demo": {
            "get": {
                "tags": [
                    "Contest",
                    "ContestDemo"
                ],
                "description": "Description api contests",
                "operationId": "83218a571d9ecfdbd3c749be13b7d94e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id cuộc thi ",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{ status: true , data : data }"
                    },
                    "404": {
                        "description": "{ status: false , message : 'Not found' }"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Nhập vào ô input dạng \"Bearer token\" <br> <b>Ví dụ</b>: <strong>Bearer 48|i93gO5EA6cBqiguiLCAiIAOiLkLoMXwoIvEnZvTn</strong>\r\n                     <br><br><hr>",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}